move test to existing file
